backend/pyproject.toml
[tool.poetry]
name = "medical-clinic"
version = "0.1.0"
description = ""
authors = ["Valera K <vk1337atliberty@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
celery = "^5.4.0"
redis = "^5.0.7"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"
flake8 = "^7.1.1"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests/*/test_*.py"]
addopts = "--cov=. --cov-report=term-missing --cov-config=pyproject.toml"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
omit = [
    "tests/*",
    "*/tests/test_*.py",
    "manage.py",
    "config/*",
    "*__init__.py",
    "migrations/*",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
